+++ CHANGELOG.md
@@ -72,5 +72,5 @@
 * OS48 - US1528 - [Presidente] Eu, diretor presidente, devo acessar a caixa de entrada para listagem de todos os processos para validar análise.
 * OS48 - US1158 - Eu, diretor técnico, devo acessar a caixa de entrada para visualizar listagem de protocolos.
 * OS48 - US1165 - Eu, diretor técnico, devo acessar a conclusão para validar a análise final do protocolo.
-* OS48 - US1528 - Eu, diretor técnico, devo acessar visualizar protocolo para visualizar os dados do protocolo.
+* OS48 - US1170 - Eu, diretor presidente, devo acessar aguardando autenticação para visualizar o resumo do protocolo.